refactor: 패키지 이동
- 설정(configuration)인 것들과 아닌 것들 분리
