+++ src/main/java/com/example/solidconnection/config/security/SecurityConfiguration.java
@@ -1,5 +1,8 @@
 package com.example.solidconnection.config.security;
 
+import com.example.solidconnection.custom.security.filter.ExceptionHandlerFilter;
+import com.example.solidconnection.custom.security.filter.JwtAuthenticationFilter;
+import com.example.solidconnection.custom.security.filter.SignOutCheckFilter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
