+++ src/main/java/com/example/solidconnection/custom/security/authentication/SiteUserAuthentication.java
@@ -1,5 +1,7 @@
-package com.example.solidconnection.config.security;
+package com.example.solidconnection.custom.security.authentication;
 
+
+import com.example.solidconnection.custom.security.userdetails.SiteUserDetails;
 
 public class SiteUserAuthentication extends JwtAuthentication {
