+++ src/main/java/com/example/solidconnection/custom/security/filter/ExceptionHandlerFilter.java
@@ -1,4 +1,4 @@
-package com.example.solidconnection.config.security;
+package com.example.solidconnection.custom.security.filter;
 
 import com.example.solidconnection.custom.exception.CustomException;
 import com.example.solidconnection.custom.response.ErrorResponse;
