+++ src/test/java/com/example/solidconnection/custom/security/filter/JwtAuthenticationFilterTest.java
@@ -1,5 +1,9 @@
-package com.example.solidconnection.config.security;
+package com.example.solidconnection.custom.security.filter;
 
+import com.example.solidconnection.config.security.JwtProperties;
+import com.example.solidconnection.custom.security.authentication.ExpiredTokenAuthentication;
+import com.example.solidconnection.custom.security.authentication.SiteUserAuthentication;
+import com.example.solidconnection.custom.security.userdetails.SiteUserDetailsService;
 import com.example.solidconnection.support.TestContainerSpringBootTest;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
